Reset seq on unexpected miss.
If I missed something, I'm not going to count it as received it.  This
isn't *quite* right, as I should keep counting forward all the time, but
I've not completely straightened out how to count up resyncs.

This doesn't actually cause it to lose any packet, just cause it to
count a telemetry window as a dropped packet after it loses a packet.